Http Server: reply object
The reply object contains the information,
that will be sent back.

It contins the functionality needed to serilize the data
to the output stream.

It also contains the constant definitions for status code,
and for error messages, that are used when returning an error reply

This file was taken from the boost example

Signed-off-by: Amnon Heiman <amnon@cloudius-systems.com>
[ penberg: cleanups ]
Signed-off-by: Pekka Enberg <penberg@cloudius-systems.com>