Http Server: handlers
The handlers are the entities that holds the specific logic
to handle a request.

All handlers inherit from handler_base.

There are a few general purpse handlers that their logic
can be reused for different scenarios.

file_handler map a url to a file, i.e. it will ignore any parameters
in the request and would always reply that file.

directory_handler is similiar to a file, but it uses a base directory
e.g. /usr/mgmt and gets a path parameter in the parameters object
it would concat the path to the base dir and would return the file.

both file and directory handle can add an optional file parser.
This is allows file that need to contain the server information,
use a parameter, and the parser would modify the result, so that
the server address apear in it.

function_handler gets a lambda expression as a parameter and uses
it to generate the result.

Signed-off-by: Amnon Heiman <amnon@cloudius-systems.com>
[ penberg: cleanups ]
Signed-off-by: Pekka Enberg <penberg@cloudius-systems.com>