Http Server: match rule
Match rule check if a given url match some logic, if it does,
it returns the handler assigned to it.

A match rule has a list of matchers, when searching
for a match, it iterate over them.
Each match returns the next position in the url,
a match would also fill the parameters object when needed.

Consider the rule /file/{path} and the url /file/etc/hosts

The match rule would hold a handler and two matcher, an
str_matcher
and a param_matcher.

It would call the first matcher with the url, it would get back
7 (the end position of the match), the parameters object is still
not changed at this point.
It would then call the second matcher that would return the
end of the string position and would fill in the parameter
object, parameters 'path' with value /etc/hosts

As this is the end of the string and there are no more matcher
the match rule would return the handler

Signed-off-by: Amnon Heiman <amnon@cloudius-systems.com>
[ penberg: cleanups ]
Signed-off-by: Pekka Enberg <penberg@cloudius-systems.com>