Http Server: routes
The routes object holds the dispatching information for
the incoming http reqeust.

It uses either an exact match or a match rules to find a
handler to handle the request based on its URL request.

If no handler is found, a page not found is returned.
If an exception is thrown during the processing of the
handler an internal server error is returned.

The matching is done in two phases, first an exact match
is search for.
For that the routes holds a map of url, if an exact match
is found, its associated handler is returned.

If there is no exact match, a matching rule is searched.
Matching rules are choose by their insertion order.
where the first rules that inserted has the highest priority.

For that, routes holds a vector of matching rules that
is search for a match.

Signed-off-by: Amnon Heiman <amnon@cloudius-systems.com>
[ penberg: cleanups ]
Signed-off-by: Pekka Enberg <penberg@cloudius-systems.com>