Http Server: Add json path support
Path information is defined in the Swagger API definition file.  This include
the url path (e.g. /os/version) the method (GET/POST) and a nickname.

A patch in the API is mapped to a path_description object and is stored in a
hash so it can be located by its nickname.

The actuall mapping of the json to c++ code is done by an extenrnal json script
during the compilation

Signed-off-by: Amnon Heiman <amnon@cloudius-systems.com>
Signed-off-by: Pekka Enberg <penberg@cloudius-systems.com>