Http Server: Expend the formatter
The formatter, format various data types to json.

This patch adds support for two additional types: unsigned long used in
the os.json swagger file and json objects.

If obj is a json object, to get a string representation of it, it should
be possible to call:

formatter::to_json(obj);

Signed-off-by: Amnon Heiman <amnon@cloudius-systems.com>
Signed-off-by: Pekka Enberg <penberg@cloudius-systems.com>