Http server: file interaction handlers complete support for file_transformer
The file_transformer is an option to modify a file while reading it and
modify part of its content.

For example, the Swagger json api, needs to contain an abosolute valid
url path. To do that a variable is placed inside the json file
'{{Host}}' and when a file is read, a transformer would replace the
{{Host}} with the host identifier taken from the request (ie.
192.168.122.89:8000)

To simplify using the transformer, it was added as an optional parameter
to the file_interaction_handler constructor and a destructor was added
to remove it.

Reviewed-by: Tomasz Grabiec <tgrabiec@gmail.com>
Signed-off-by: Amnon Heiman <amnon@cloudius-systems.com>
Signed-off-by: Pekka Enberg <penberg@cloudius-systems.com>