Testing: Adding a testing script for httpserver
This is a testing script for the httpserver, it now test the os api, but
adds infrastructure to test additional APIs when they will be added.

The test script starts an image and run the httpserver on it it later
tests each of the os API method calls and end with shuting down the
image with the shutdown API call.

Signed-off-by: Amnon Heiman <amnon@cloudius-systems.com>
Signed-off-by: Pekka Enberg <penberg@cloudius-systems.com>