Http Server: Remove copy constructor and assignment from json base
In the implementation of the jvm API there is a used of a list of Json
object, that cause sometimes Json objects to be copied construct.

To support Json object copying, the base copy constructor does not needs
to do anything as the copying is done in the inherit class. To enforce
that, the copy constructor and assignment operator are deleted from the
json base class.

The logic for copying an object is added by the code generation.

Signed-off-by: Amnon Heiman <amnon@cloudius-systems.com>