Http Server: Add not found exception to handler
Sometimes it is required that a handler will return a 404, for example
the handler that interact with the jvm will return a 404 if the jvm is
not available.

To support that the handler can throw a not_found exception that will be
caught by the routes that in response will return a 404 to the request.

In addition the set_header method was change so that a handler can use
it to return status different then an OK.

Signed-off-by: Amnon Heiman <amnon@cloudius-systems.com>