Swagger API: Adding the file API
The definition of the file API are based on the hadoop file-system API
definitions found under:

http://hadoop.apache.org/docs/stable/hadoop-project-dist/hadoop-hdfs/WebHDFS.html

The hadoop api is based on defining the path to the resource (file) and
an operation. This could be sometimes redundant like in the case of
DELETE and it cannot be as document as clearly as operation that are
part of the path, but we choose to take their suggestion as it is
already known.

The main difference between the hadoop file system API in this API is
file uploading. The API uses the more common API definition of POST a
file, instead of the hadoop file system that mimic a file operating
system with open, append and close.

Note that the approach of the standard POST file, is supported by
swagger and allows a simple file-picker API to be used, as apposed to
the hadoop file-system approach. It is also means that the http-server
can remain stateless.

Here are a few example taken from hadoop file system api:

Reading from a file:
GET: http://<HOST>:<PORT>/file/<PATH>?op=OPEN[&offset=<LONG>][&length=<LONG>][&buffersize=<INT>]

Creating a directory:
PUT: http://<HOST>:<PORT>/file/<PATH>?op=MKDIRS[&permission=<OCTAL>]

Rename a File/Directory:
PUT: http://<HOST>:<PORT>/file/<PATH>?op=RENAME&destination=<PATH>

Delete a File/Directory:
DELETE: http://<host>:<port>/file/<path>?op=DELETE[&recursive=<true|false>]

Status of a File/Directory:
GET: http://<HOST>:<PORT>/file/<PATH>?op=GETFILESTATUS

Signed-off-by: Amnon Heiman <amnon@cloudius-systems.com>