Shell: Stop building an uberjar for CRaSH
This patch removes the complicated build we had for CRaSH which created an
uberjar. Dependencies are copied using the build module's FileMap.

The crash jar is also placed in its own classpath, before the other library
jars because our class overrides one of CRaSH's internal classes which is
loaded with ServiceLoader.

Signed-off-by: Or Cohen <orc@fewbytes.com>
Signed-off-by: Pekka Enberg <penberg@cloudius-systems.com>