Shell: Upgrade to crash-1.3.0-beta15
This version comes with the option to specify multiple auth plugins. As
a result, I've selected a one-size-fits-all kind of value.

I also moved our crash.properties file to its own folder and updated the config
path in the main CRaSH command. Before that, CRaSH was picking up the config
file that is bundled in CRaSH jars.

Signed-off-by: Or Cohen <orc@fewbytes.com>
Signed-off-by: Pekka Enberg <penberg@cloudius-systems.com>