Move REST API spec to a separate project
REST API is defined using JSON files base on Swagger spec.
They are currently located at a dir under mgmt project.

This force anyone who wish to use the API specs to pull the entire mgmt project.
Users will need only the JSON files to build, for example a Ruby client.

This is the second patch of two, which will move mgmt/api to a new osv-rest-api repository.
The first patch added the files to the new repository
In this patch mgmt/api directory is replaced with a submodule.
In addition, Make file has been update to do 'git submodule update' in case the API sub module is not there.

Signed-off-by: Tzach Livyatan <tzach@cloudius-systems.com>
Signed-off-by: Pekka Enberg <penberg@cloudius-systems.com>