Http Server: Add enum support to code generation
When a parameter is declare as enum in the swagger definition file, the
code generation will create an enum internal namespace to the operation.

To the set of values a NUM_ITEMS value is added as the last value, so
if an enum op is declared as:
"enum":["MKDIRS", "RENAME"] in an operation putFile
The declared enum would be:

namespace ns_putFile {

enum class op {
MKDIRS, RENAME, NUM_ITEMS
};

For each enum a translation from string is created for it. The
translation function return the enum value if it is found or NUM_ITEMS
if it is not found, similar to string method that returns npos.

Signed-off-by: Amnon Heiman <amnon@cloudius-systems.com>
Signed-off-by: Avi Kivity <avi@cloudius-systems.com>