Http Server: Clean up in the lambda expression handlers
To simplify the use of lambda expression as handlers, the function type
was define, and additional function was added, a one that gets not only
the request but also the reply, this will allow using
lambda expression in more complicated scenarios then before.

The function handler was modified to use the more general function and a
temporary function is used to do the translation.

Signed-off-by: Amnon Heiman <amnon@cloudius-systems.com>
Signed-off-by: Avi Kivity <avi@cloudius-systems.com>