Http Server: handler to support mandatory parameters
Mandatory parameters are query parameters that the handler expect to
exists before starting the handler logic.

The handler now holds a list of mandatory parameter, that will be verify
before the handler is called.

The handler logic can safely assume all the mandatory parameters are
there and doesn't need to verify it.
If the parameters are missing a 400 will be return with the appropriate
error message.

The code generator will automatically adds any mandatory parameter from
the Swagger definition file. But it is also possible to add additional
parameters.

if h is the handler calling:
h.mandatory("op");

will add the query parameter "op" as a mandatory parameter

Signed-off-by: Amnon Heiman <amnon@cloudius-systems.com>
Signed-off-by: Avi Kivity <avi@cloudius-systems.com>