Http server: Adding path holder
Path holder is an object that simplifies setting a handler to an api
define in the Swagger definition file.

The code generation will set a path holder object for each of the api
calls. For example to the api action "getFile" the following will be
define:

static const path_holder getFile("getFile");

getFile can now be used to set a handler to the operation. So if h is a
handler you can do:

getFile.set_handler(h);

It also support lamda expression, so if f is a request_function (as
define in the handlers.hh file) you can do:

getFile.set_handler(f,"json");

With the new API there is no need to access the routes object in most
cases.

Signed-off-by: Amnon Heiman <amnon@cloudius-systems.com>
Signed-off-by: Avi Kivity <avi@cloudius-systems.com>