mgmt/httpserver: Add more include directories
httpserver's Makefile lists some OSv include directories so we can use OSv
APIs in our implementation of some of the requests. However, we did not
add enough include directories to allow including some OSv headers, such
as <osv/sched.hh>.

This patch adds to the include path ../../arch/$(arch), which is needed
for <arch.hh>, and ../.., needed for <drivers/clockevent.hh>. Both are
needed to be able to include <osv/sched.hh>.

To be honest, OSv's C++ API situation is a bit of a mess - there are too
many different directories of header files, and the API headers like
sched.hh include a huge number of other headers to define all the class
fields which are unfortunately all exposed. It's also not clear what is
the official policy of mixing Linux and OSv header files. But before we
solve these problems, this patch is an interim fix.

Signed-off-by: Nadav Har'El <nyh@cloudius-systems.com>
Signed-off-by: Pekka Enberg <penberg@cloudius-systems.com>