REST API: Add /os/thread request
Add a /os/threads request that can be used to implement a top command
(in a separate patch). The request returns the list of all threads,
and for each thread specifies its numeric id, it's name, and its total
CPU time usage. The list is also accompanied by the current time (this
is useful when "top" is run over a slow network).

All times are printed in millisecond resolution, which I deemed enough
for the purpose of this API (we have nanosecond resolution, but it will
needlessly increase the size of the output).

In the future, additional fields such as a thread's CPU, priority, pinned
status, state (running, waiting, etc.), stack size, etc., can be added.
The fields I added now are the minimum selection of fields necessary for
a useful "top" that helps finding CPU-hogging threads.

Signed-off-by: Nadav Har'El <nyh@cloudius-systems.com>